file-ingestion-service/pyproject.toml
[tool.poetry]
name = "file-ingestion-service"
version = "0.1.0"
description = ""
authors = ["Dimas Tri Mustakim <test31122001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.12"
langchain-community = "^0.2.11"
langchain-experimental = "^0.0.64"
python-dotenv = "^1.0.1"
neo4j = "^5.23.1"
langchain-google-vertexai = "^1.0.8"
pymupdf = "^1.24.9"
tiktoken = "^0.7.0"
anthropic = {extras = ["vertex"], version = "^0.32.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
